% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.lctm.escc-package.R
\docType{package}
\name{quick.lctm.escc-package}
\alias{quick.lctm.escc}
\alias{quick.lctm.escc-package}
\title{quick.lctm.escc: Latent Class Trajectory Modeling for 7 Nutritional Markers}
\description{
Fits Bayesian latent class trajectory models (LCTM) on 7 markers
(ALB, TC, LYM, CRP, SMI, SFI, VFI) at multiple time points to
identify nutritional trajectory phenotypes in cancer patients.
}
\details{
Two-level input design (7 markers only; no baseline covariates):
\itemize{
  \item L1 - \code{\link{fit_lctm_full}}: Full 5-timepoint LCTM
  \item L2 - \code{\link{fit_lctm_partial}}: 2-4 timepoints with MCMC imputation
}

Includes pre-trained v8 posterior parameters (\code{\link{lctm_v8_params}})
for offline prediction of new patients without re-running MCMC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Tanly78/Quick_LCTM_for_ESCC}
  \item Report bugs at \url{https://github.com/Tanly78/Quick_LCTM_for_ESCC/issues}
}

}
\author{
Anonymous \email{package@example.com} (ORCID: 0009-0000-6575-5277)
}
\keyword{package}
